+++ Kyoo.Common/Models/GenreLink.cs
@@ -2,9 +2,9 @@ namespace Kyoo.Models
 {
 	public class GenreLink
 	{
-		public long ShowID { get; set; }
+		public int ShowID { get; set; }
 		public virtual Show Show { get; set; }
-		public long GenreID { get; set; }
+		public int GenreID { get; set; }
 		public virtual Genre Genre { get; set; }
 
 		public GenreLink() {}
